# [luci-interpreter] Introduce MemoryManager Interface and SimpleMemoryManager