This commit introduces memory manager interface and simple memory manager realization. This is first step of merging changes from pull request #7408 

issue #7355

ONE-DCO-1.0-Signed-off-by: Artem Balyshev a.balyshev@partner.samsung.com