@seanshpark 2021-08-30
Q) Is luci-interpreter default memory manager going to be changed to something else?
(I didn't review the draft code so don't know much)
Why I am asking is that luci-interpreter is used as test framework in the CI and I assume changing something may affect the performance. If it gets faster that will be a good news but if it gets slower, our CI maybe unhappy.

CC @binarman