@binarman 2021-08-30
@seanshpark 
> Is luci-interpreter default memory manager going to be changed to something else?

In short: no, there should be no changes in changes in default memory allocations order/time.

In detail: Currently there are no such separate entity as memory manger in interpreter. All allocations/deallocations happen dynamically and algorithm **embedded in core interpreter**.
In order to generalize code structure, default allocation strategy is moved to SimpleMemoryManager.

I understand that it is not obvious that everything is the same, so I am thinking on testing methods right now.