@BalyshevArtem 2021-09-06
Result of evaluation inception_v3, inception_v4, mobilenet_v1_224, mobilenet_v2_224 and resnet_v2_101_299 with memory manager and without.

[resnet_v2_101_299_without_mm.txt](https://github.com/Samsung/ONE/files/7116482/resnet_v2_101_299_without_mm.txt)
[resnet_v2_101_299_with_mm.txt](https://github.com/Samsung/ONE/files/7116483/resnet_v2_101_299_with_mm.txt)
[mobilenet_v2_without_mm.txt](https://github.com/Samsung/ONE/files/7116484/mobilenet_v2_without_mm.txt)
[mobilenet_v2_with_mm.txt](https://github.com/Samsung/ONE/files/7116485/mobilenet_v2_with_mm.txt)
[mobilenet_v1_without_mm.txt](https://github.com/Samsung/ONE/files/7116486/mobilenet_v1_without_mm.txt)
[mobilenet_v1_with_mm.txt](https://github.com/Samsung/ONE/files/7116487/mobilenet_v1_with_mm.txt)
[inception_v4_without_mm.txt](https://github.com/Samsung/ONE/files/7116488/inception_v4_without_mm.txt)
[inception_v4_with_mm.txt](https://github.com/Samsung/ONE/files/7116489/inception_v4_with_mm.txt)
[inception_v3_without_mm.txt](https://github.com/Samsung/ONE/files/7116491/inception_v3_without_mm.txt)
[inception_v3_with_mm.txt](https://github.com/Samsung/ONE/files/7116492/inception_v3_with_mm.txt)