@binarman 2021-09-07
FYI @seanshpark 

> Q) Is luci-interpreter default memory manager going to be changed to something else?

**Summary**

Tables to compare results with and without new memory schema.
Based on results provided by @BalyshevArtem , models taken from TF hosted models page.

Execution time

Model name | Current implementation (Sec) | With memory manager (Sec)
--- | --- | ---
resnet 101 | 6.52 | 6.48
mobilenet v1 | 0.54 | 0.54
mobilenet v2 | 0.29 | 0.30
inception v3 | 1.77 | 1.74
inception v4 | 4.02 | 4.00

Memory consumption

Model name | Current implementation (kbytes) | With memory manager (kbytes)
--- | --- | ---
resnet 101 | 530216 | 530052
mobilenet v1 | 56500 | 56508
mobilenet v2 | 48032 | 47864
inception v3 | 286648 | 286592
inception v4 | 507704 | 507552

I don't see any performance impact so far. Differences are probably result of imperfection of the benchmark environment (it's a developer's PC without any special preparations).