@m-bronnikov 2021-09-09
Could you please change this PR title with accordance with commit title (add mention of `SimpleMemoryManager` introduction)?