@mhs4670go 2021-09-17
@BalyshevArtem 
```cxx
auto *data = new uint8_t[num_elements * element_size];
  tensor.set_data_buffer(data);
```

Hi, the static analysis tool said resource leak can be occurred. Will this be resolved another PR?