@BalyshevArtem 2021-09-17
@mhs4670go 
Hi! The memory should not leak because the release method is called for each allocated memory for the tensor. To check the absence of leaks, we conducted tests using  leak sanitizer and address sanitizer. And everything was okay :)